requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.0
pydantic==2.4.2